refactor: 검색어 추천 모달 커스텀 hook으로 모듈화
